requirements.txt
asgiref==3.7.2
crispy-bootstrap5==2023.10
Django==5.0.3
django-crispy-forms==2.1
djangorestframework==3.15.2
fontawesomefree==6.5.1
gunicorn==21.2.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.10.0
